+++ app/anon/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import AnonChatScreen from "@/components/screens/AnonChatScreen";
+
+export default function Home() {
+  return <AnonChatScreen />;
+}
