+++ app/api/anon/route.ts
@@ -1,60 +1,17 @@
 import { NextResponse } from "next/server";
 import { MembershipVerifier } from "@personaelabs/spartan-ecdsa";
-import prisma from "@/lib/prisma";
 import { Telegraf } from "telegraf";
 import path from "path";
 
 const telegramBot = new Telegraf(process.env.TELEGRAM_BOT_TOKEN!);
 const TELEGRAM_TEST_CHAT_ID = -4031859798;
-const emojiMap: { [key: string]: string } = {
-  "robot.png": "🤖",
-  "invader.png": "👾",
-  "ninja.png": "🥷",
-  "turtle.png": "🐢",
-  "pizza.png": "🍕",
-  "pinata.png": "🪅",
-  "unicorn.png": "🦄",
-  "mushroom.png": "🍄",
-  "soccer.png": "⚽️",
-  "pumpkin.png": "🎃",
-  "avocado.png": "🥑",
-  "nerd.png": "🤓",
-  "ghost.png": "👻",
-  "cowboy.png": "🤠",
-  "alien.png": "👽",
-  "icecream.png": "🍦",
-  "fairy.png": "🧚",
-  "butterfly.png": "🦋",
-  "bubbles.png": "🧼",
-  "dolphin.png": "🐬",
-  "angry-cat.png": "😾",
-  "baby-angel.png": "👼",
-  "blue-swirl.png": "🌀",
-  "boba.png": "🧋",
-  "caterpillar.png": "🐛",
-  "clown.png": "🤡",
-  "cookie.png": "🍪",
-  "crystal-ball.png": "🔮",
-  "dumpling.png": "🥟",
-  "eyes.png": "👀",
-  "fingers-crossed.png": "🤞",
-  "handshake.png": "🤝",
-  "head-exploding.png": "🤯",
-  "lollipop.png": "🍭",
-  "magic-wand.png": "🪄",
-  "palette.png": "🎨",
-  "poop.png": "💩",
-  "skull.png": "💀",
-  "sloth.png": "🦥",
-  "squirrel.png": "🐿",
-};
 
 /**
- * POST /api/telegram
+ * POST /api/anon
  * Posts a pseudonymous chat message which is verified and broadcasted to the Telegram chat bot.
  * @param {object} request - The request object. Consists of the following parameters encoded as JSON.
+ * @param {string} request.pseudonym - The pseudonym the user is posting as.
  * @param {string} request.message - The message the user is posting.
- * @param {string} request.sigmoji - The Sigmoji the user is posting as. Currently represents the emojiImg of the Sigmoji.
  * @param {string} request.serializedZKP - The serialized zero-knowledge proof provided by the user.
  * @returns A JSON response with a success boolean indicating if the chat message was posted.
  */
@@ -62,7 +19,7 @@ export async function POST(request: Request) {
   try {
     const bodyArrayBuffer = await request.arrayBuffer();
     const bodyString = new TextDecoder().decode(bodyArrayBuffer);
-    const { message, sigmoji, serializedZKP } = JSON.parse(bodyString);
+    const { pseudonym, message, serializedZKP } = JSON.parse(bodyString);
 
     // can't use deserializeSigmojiZKP due to some Next.js problem
     const parsedZKP = JSON.parse(serializedZKP);
@@ -76,15 +33,8 @@ export async function POST(request: Request) {
       throw new Error("Proof failed verification");
     }
 
-    if (!emojiMap.hasOwnProperty(sigmoji)) {
-      throw new Error("Sigmoji not found");
-    }
-
-    const emoji = emojiMap[sigmoji];
-    const fullMessage = `${emoji}: ${message}`;
+    const fullMessage = `A Sigmoji holder (${pseudonym}): ${message}`;
     telegramBot.telegram.sendMessage(TELEGRAM_TEST_CHAT_ID, fullMessage);
-
-    await addChatLog({ message, sigmoji });
 
     return new NextResponse(undefined, {
       status: 200,
@@ -120,21 +70,4 @@ async function verifyProof(zkp: {
   await verifier.initWasm();
 
   return await verifier.verify(zkp.proof, zkp.publicInputSer);
-}
-
-/**
- * Adds a new entry to the chat log.
- * @param logEntry - The entry to add.
- * @param logEntry.message - The message that was sent.
- * @param logEntry.sigmoji - The Sigmoji the user is posting as. Currently represents the emojiImg of the Sigmoji.
- */
-async function addChatLog(logEntry: { message: string; sigmoji: string }) {
-  try {
-    const newEntry = await prisma.chatLog.create({
-      data: logEntry,
-    });
-    console.log("New chat log entry:", newEntry);
-  } catch (error) {
-    console.error("Error adding chat log entry:", error);
-  }
 }
