+++ components/screens/AnonChatScreen.tsx
@@ -4,7 +4,6 @@ import Footer from "@/components/shared/Footer";
 import { TextArea } from "@/components/shared/TextArea";
 import { PrimaryLargeButton } from "@/components/shared/Buttons";
 import { CourierPrimeBase, PrimaryFontH1 } from "@/components/core";
-import Image from "next/image";
 import { useEffect, useState } from "react";
 import { Sigmoji } from "@/lib/types";
 import { loadSigmojis } from "@/lib/localStorage";
@@ -14,6 +13,7 @@ import {
   initWasm,
   setupTree,
 } from "@/lib/zkProving";
+import { Input } from "../shared/Input";
 
 enum ChatDisplayState {
   LOADING,
@@ -24,9 +24,8 @@ enum ChatDisplayState {
 
 export default function ChatScreen() {
   const [wasm, setWasm] = useState<ProverWasm>();
-  const [sigmojiArr, setSigmojiArr] = useState<Sigmoji[]>([]);
-  const [selectedSigmoji, setSelectedSigmoji] = useState<string>();
-  const [message, setMessage] = useState("");
+  const [pseudonym, setPseudonym] = useState<string>("");
+  const [message, setMessage] = useState<string>("");
   const [displayState, setDisplayState] = useState<ChatDisplayState>(
     ChatDisplayState.LOADING
   );
@@ -35,13 +34,6 @@ export default function ChatScreen() {
     async function setup() {
       if (!wasm) {
         setWasm(await initWasm());
-      }
-
-      const arr = await loadSigmojis();
-      setSigmojiArr(arr);
-
-      if (arr.length > 0) {
-        setSelectedSigmoji(arr[0].emojiImg);
       }
     }
 
@@ -53,10 +45,6 @@ export default function ChatScreen() {
       setDisplayState(ChatDisplayState.READY);
     }
   }, [wasm]);
-
-  const onSelectSigmoji = (emojiImg: string) => {
-    setSelectedSigmoji(emojiImg);
-  };
 
   const generateProofForSigmoji = async (
     sigmoji: Sigmoji
@@ -72,33 +60,38 @@ export default function ChatScreen() {
   };
 
   const onSubmit = async () => {
+    if (!pseudonym) {
+      alert("Please select a pseudonym.");
+      return;
+    }
+
     if (!message) {
       alert("Please enter a message.");
       return;
     }
 
-    let sigmoji = sigmojiArr.find(
-      (sigmoji) => sigmoji.emojiImg === selectedSigmoji
-    );
-    if (!sigmoji) {
-      alert("Invalid Sigmoji selected.");
+    const sigmojiArr = await loadSigmojis();
+    if (sigmojiArr.length === 0) {
+      alert("You must have a Sigmoji to use this chat.");
       return;
     }
 
+    // Just use the first sigmoji for now
+    let sigmoji = sigmojiArr[0];
     if (!sigmoji.ZKP) {
       setDisplayState(ChatDisplayState.PROVING);
       sigmoji = await generateProofForSigmoji(sigmoji);
     }
 
     setDisplayState(ChatDisplayState.SUBMITTING);
-    await fetch("/api/chat", {
+    await fetch("/api/anon", {
       method: "POST",
       headers: {
         "Content-Type": "application/json",
       },
       body: JSON.stringify({
+        pseudonym: pseudonym,
         message: message,
-        sigmoji: sigmoji.emojiImg,
         serializedZKP: sigmoji.ZKP,
       }),
     }).then(async (response) => {
@@ -134,45 +127,17 @@ export default function ChatScreen() {
       <MainHeader />
       <ChatContainer>
         <PrimaryFontH1 style={{ color: "var(--woodsmoke-100)" }}>
-          Chat
+          Anon Chat
         </PrimaryFontH1>
         <CourierPrimeBase style={{ maxWidth: "90%" }}>
           Chat pseudonymously with other Sigmoji holders! Messages will be sent
           to the Sigmoji Telegram group.
         </CourierPrimeBase>
-        {selectedSigmoji && (
-          <SelectionContainer>
-            <CourierPrimeBase style={{ color: "var(--woodsmoke-100)" }}>
-              Select a Sigmoji to chat as:
-            </CourierPrimeBase>
-            <SigmojiContainer>
-              {sigmojiArr.map((sigmoji, index) => (
-                <button
-                  key={index}
-                  onClick={() => onSelectSigmoji(sigmoji.emojiImg)}
-                >
-                  <Image
-                    src={`/emoji-photo/${sigmoji.emojiImg}`}
-                    width="24"
-                    height="24"
-                    alt="emoji"
-                  />
-                </button>
-              ))}
-            </SigmojiContainer>
-            <CurrentSelectionContainer>
-              <CourierPrimeBase style={{ color: "var(--woodsmoke-100)" }}>
-                You are currently chatting as:
-              </CourierPrimeBase>
-              <Image
-                src={`/emoji-photo/${selectedSigmoji}`}
-                width="24"
-                height="24"
-                alt="emoji"
-              />
-            </CurrentSelectionContainer>
-          </SelectionContainer>
-        )}
+        <CourierPrimeBase style={{ maxWidth: "90%" }}>
+          You will only be defined by the pseudonym you select, but you must
+          have a Sigmoji to use this chat.
+        </CourierPrimeBase>
+        <Input header="Pseudonym" value={pseudonym} setValue={setPseudonym} />
         <TextArea header="Message" value={message} setValue={setMessage} />
         <PrimaryLargeButton disabled={!wasm} onClick={onSubmit}>
           {getDisplayText()}
@@ -193,25 +158,4 @@ const ChatContainer = styled.div`
   text-align: center;
   align-items: center;
   align-self: stretch;
-`;
-
-const SelectionContainer = styled.div`
-  display: flex;
-  gap: 8px;
-  flex-direction: column;
-  align-items: center;
-`;
-
-const SigmojiContainer = styled.div`
-  display: flex;
-  gap: 16px;
-  flex-direction: row;
-  align-items: center;
-`;
-
-const CurrentSelectionContainer = styled.div`
-  display: flex;
-  gap: 4px;
-  flex-direction: row;
-  align-items: center;
 `;
